feature: DynamicObject
