+++ src/components/Presentation/SlideCanvasLayout/Object/Vertex/index.jsx
